move customenavbar to /widgets
